Day2Part2 Linear Update
Ditching the recursive approach for a simpler, iterative appraoch. I iterate through all the windows of size 3. When 1 doesn't work we'll attempt to remove all 3 combinations of items and see if the window now works. If it doesn't then the numbers are unsafe. If one of them works, then we move on to the next window. If that window doesn't work, then we abort because we've already fixed one window.

I think this approach will work but time will tell